+++ EL_FitsFile/src/program/EL_FitsIO_ReadStructure.cpp
@@ -92,7 +92,7 @@ public:
         if (shape.size() > 0) {
           std::ostringstream oss;
           std::copy(shape.begin(), shape.end() - 1, std::ostream_iterator<int>(oss, " x "));
-          oss << shape.back();
+          oss << shape.indices.back();
           logger.info() << "  Image HDU:";
           logger.info() << "    Type: " << readBitpixName(hdu.as<ImageHdu>());
           logger.info() << "    Shape: " << oss.str() << " px";